+++ e2e/test-resources/sample-repo/environment_variables/JavaEnv.java
@@ -1,0 +1,5 @@
+package environment_variables;
+
+public class JavaEnv {
+  public static final String value = "value";
+}