+++ e2e/test-resources/sample-repo/environment_variables/JavaTest.java
@@ -1,0 +1,5 @@
+package environment_variables;
+
+public class JavaTest {
+  public static final String value = "value";
+}