+++ server/src/main/java/org/jetbrains/bsp/bazel/server/sync/BazelProjectMapper.kt
@@ -5,9 +5,11 @@ import kotlinx.coroutines.flow.filterNot
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.toList
 import kotlinx.coroutines.runBlocking
+import org.jetbrains.bsp.bazel.info.BspTargetInfo
 import org.jetbrains.bsp.bazel.info.BspTargetInfo.FileLocation
 import org.jetbrains.bsp.bazel.info.BspTargetInfo.TargetInfo
 import org.jetbrains.bsp.bazel.server.sync.dependencytree.DependencyTree
+import org.jetbrains.bsp.bazel.server.sync.languages.LanguageData
 import org.jetbrains.bsp.bazel.server.sync.languages.LanguagePlugin
 import org.jetbrains.bsp.bazel.server.sync.languages.LanguagePluginsService
 import org.jetbrains.bsp.bazel.server.sync.model.Label
@@ -18,6 +20,8 @@ import org.jetbrains.bsp.bazel.server.sync.model.SourceSet
 import org.jetbrains.bsp.bazel.server.sync.model.Tag
 import org.jetbrains.bsp.bazel.workspacecontext.WorkspaceContext
 import java.net.URI
+import java.util.stream.Collector
+import java.util.stream.Collectors
 
 class BazelProjectMapper(
     private val languagePluginsService: LanguagePluginsService,
@@ -74,6 +78,7 @@ class BazelProjectMapper(
         val resources = resolveResources(target)
         val languageData = languagePlugin.resolveModule(target)
         val sourceDependencies = languagePlugin.dependencySources(target, dependencyTree)
+        val environment = environmentItem(target)
         return Module(
             label = label,
             isSynthetic = false,
@@ -84,7 +89,8 @@ class BazelProjectMapper(
             sourceSet = sourceSet,
             resources = resources,
             sourceDependencies = sourceDependencies,
-            languageData = languageData
+            languageData = languageData,
+            environmentVariables = environment
         )
     }
 
@@ -129,4 +135,14 @@ class BazelProjectMapper(
         with(module) {
             (sourceSet.sources + resources).map { Pair(it, label) }
         }
+
+    private fun environmentItem(target: TargetInfo): Map<String, String> {
+        val inheritedEnvs = collectInheritedEnvs(target)
+        val targetEnv = target.envMap
+        return inheritedEnvs + targetEnv
+    }
+
+    private fun collectInheritedEnvs(targetInfo: TargetInfo): Map<String, String> =
+         targetInfo.envInheritList.associateWith { System.getenv(it)}
+
 }