+++ server/src/main/java/org/jetbrains/bsp/bazel/server/sync/model/Module.kt
@@ -15,4 +15,5 @@ data class Module(
     val resources: Set<URI>,
     val sourceDependencies: Set<URI>,
     val languageData: LanguageData?,
+    val environmentVariables: Map<String, String>
 )