+++ server/src/test/java/org/jetbrains/bsp/bazel/server/bloop/BspModuleExporterTest.kt
@@ -50,9 +50,10 @@ class BspModuleExporterTest {
                     emptyList(),
                     listOf(baseDirectory.resolve("cp1").toUri()),
                     emptyList(),
-                    emptyList()
+                    emptyList(),
                 )
-            )
+            ),
+            hashMapOf()
         )
         val project = Project(
             baseDirectory.toUri(), listOf(bspModule), emptyMap()